Decouple subjects from users Remove users from the subjects data. Read user subject comments from the discussionComments helper.
