+++ src/helpers/subjects.js
@@ -16,7 +16,6 @@ function parseRow({ ouroboros_subject_id }) {
 
 async function fetchSubjects() {
   const { name } = await project;
-  const users = await awaitUsers;
 
   const parseData = new Promise((resolve, reject) => {
     fs.createReadStream(path.resolve(__dirname, '../../.data', `${name}_subject_ids.csv`))
@@ -35,25 +34,6 @@ async function fetchSubjects() {
 
   for (subject of APIsubjects) {
     subjects[subject.zooniverse_id] = subject;
-    for (comment of subject.discussion.comments) {
-      const author = store.users[comment.user_name];
-      const focus = {
-        location: subject.location,
-        zooniverse_id: subject.zooniverse_id
-      }
-      const commentExists = author && author.subjects.find(userSubject => userSubject.comment._id === comment._id);
-      if (author && !commentExists) {
-        author.subjects.push({ comment, focus });
-      }
-    }
-  }
-
-  for (user of Object.values(store.users)) {
-    user.subjects.sort(function sortSubjectComments(subject1, subject2) {
-      const date1 = new Date(subject1.comment.created_at);
-      const date2 = new Date(subject2.comment.created_at);
-      return date2 - date1;
-    });
   }
 
   return subjects;
